데미지 처리 추가
1. 공격이 클릭마다(애니메이션 사용중에도) 발생하는 오류 해결
2. enemy base class를 기반으로 데미지 처리, 체력 적용

+++ Source/DreamWorld/FrameWork/ActorComponent/AC_AttackManager.h
@@ -21,7 +21,7 @@ public:
 	* 인자 : 공격 시도자, 데미지
 	* 공격 범위 체크 및 데미지 처리
 	*/
-	void CallAttemptAttack(TObjectPtr<AActor> in_Player, uint32 in_Damage);
+	void CallAttemptAttack(TObjectPtr<AActor> in_Player, float in_Damage);
 
 public:
 	/** Getter */
